jobrelay/requirements.jobrelay.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
duckdb>=1.0.0
pandas>=2.0.0
pyarrow>=14.0.0
gcsfs>=2023.12.0
google-cloud-storage>=2.16.0
python-dateutil>=2.8.2
pydantic-settings>=2.2.1
httpx>=0.27.0
pytz>=2024.1
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
psycopg[binary,pool]>=3.1
